avoid unsafe capture of this in MultiFrameCodec
