+++ lib/ui/fixtures/ui_test.dart
@@ -3,4 +3,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:ui';
+
 void main() {}
+
+@pragma('vm:entry-point')
+void frameCallback(FrameInfo info) {
+  print('called back');
+}
