+++ testing/BUILD.gn
@@ -47,6 +47,7 @@ source_set("dart") {
     "elf_loader.h",
     "test_dart_native_resolver.cc",
     "test_dart_native_resolver.h",
+    "dart_isolate_runner.h",
   ]
 
   public_deps = [
